moved out nativeTest into utils/testing
